requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
black==25.1.0
click==8.1.8
discord.py==2.3.2
frozenlist==1.4.1
idna==3.10
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.0
pytz==2024.1
ruff==0.9.7
typing_extensions==4.11.0
yarl==1.9.4
